+++ src/azure-firewall/azext_firewall/__init__.py
@@ -13,9 +13,8 @@ class AzureFirewallCommandsLoader(AzCommandsLoader):
 
     def __init__(self, cli_ctx=None):
         from azure.cli.core.commands import CliCommandType
-        from .profiles import CUSTOM_FIREWALL, CUSTOM_FIREWALL_POLICY
-        register_resource_type('latest', CUSTOM_FIREWALL, '2019-11-01')
-        register_resource_type('latest', CUSTOM_FIREWALL_POLICY, '2020-05-01')
+        from .profiles import CUSTOM_FIREWALL
+        register_resource_type('latest', CUSTOM_FIREWALL, '2020-05-01')
 
         super(AzureFirewallCommandsLoader, self).__init__(
             cli_ctx=cli_ctx,