+++ src/ai-examples/HISTORY.rst
@@ -3,6 +3,15 @@
 Release History
 ===============
 
+0.2.2
+++++++
+* Fix bug causing queries to fail when not logged in and when telemetry is disabled
+
+0.2.1
+++++++
+* Change telemetry collection
+* Improve README
+
 0.2.0
 ++++++
 * Add AI generated examples to all group and command "--help"