+++ src/aks-preview/HISTORY.md
@@ -2,6 +2,25 @@
 
 Release History
 ===============
+0.4.56
++++++
+* Support "--enable-aad" for "az aks update" to update an existing RBAC-enabled non-AAD cluster to the new AKS-managed AAD experience
+
+0.4.55
++++++
+* Add "--enable-azure-rbac" for enabling Azure RBAC for Kubernetes authorization
+
+0.4.54
++++++
+* Support "--enable-aad" for "az aks update" to update an existing AAD-Integrated cluster to the new AKS-managed AAD experience
+
+0.4.53
++++++
+* Add --ppg for "az aks create" and "az aks nodepool add"
+
+0.4.52
++++++
+* Add --uptime-sla for az aks update
 
 0.4.51
 +++++