+++ src/spring-cloud/HISTORY.md
@@ -1,6 +1,15 @@
 Release History
 ===============
 
+0.3.1
+-----
+* Remove azure-storage-blob dependency
+
+0.3.0
+-----
+* Enable distributed tracing by default when creating the service
+* Enable to update tags and distributed tracing settings by using "az spring-cloud update"
+
 0.2.6
 -----
 * Fix required sku issue